% Generated by roxygen2 (4.0.2): do not edit by hand
\name{svar}
\alias{svar}
\title{Test documentation}
\usage{
svar(ts, p, lambda = NULL, cv.lambda = TRUE, verbose = TRUE,
  method = c("lm", "ols", "olsl2", "l1.ls", "l1.ll"), type = c("none",
  "const", "trend", "both"), epsilon.min = 0.01)
}
\description{
Test documentation
}

